resolved the long output error in task 14

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -10,22 +10,22 @@ int main(void)
 {
 	int i;
 	/* initializing the first two terms*/
-	int term1 = 1, term2 = 2;
+	long term1 = 1, term2 = 2;
 	/*initialize the next term*/
-	int next_term = term1 + term2;
+	long next_term = term1 + term2;
 
 	/*printing the first two terms*/
-	printf("%d, %d, ", term1, term2);
+	printf("%ld, %ld, ", term1, term2);
 	/*printing 3rd to 50th term*/
 	for (i = 3; i <= 50; ++i)
 	{
 		if (i != 50)
 		{
-			printf("%d, ", next_term);
+			printf("%ld, ", next_term);
 		}
 		else
 		{
-			printf("%d\n", next_term);
+			printf("%ld\n", next_term);
 		}
 		term1 = term2;
 		term2 = next_term;
